document_neo4j_etl/pyproject.toml
[project]
name = "document_neo4j_etl"
version = "0.1"
dependencies = [
    "neo4j==5.14.1",
    "retry==0.9.2",
    "pypdf2==3.0.1",
    "openai==0.27.0",
    "networkx==2.8.8",
    "python-dotenv==1.0.0"
]

[project.optional-dependencies]
dev = ["black", "flake8"]
